refactor: relocate pop-api specifics to api pallet
